requirements.txt
sqlalchemy
pydantic
fastapi
python-dotenv
uvicorn
psycopg2-binary
poetry
